requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
beautifulsoup4==4.12.2
requests==2.31.0
apscheduler>=3.11.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-telegram-bot==20.7
myjdapi
